[ARCH-P4] Move query receipts behind application/query
